coprocessor/endpoint:adjust test
